config/action/MessageAction: change type of values to Handlebars
So there is almost no difference between this and action::MessageAction.
The plan is to change action::MessageAction to contain a response sender
and a MessageAction value (now, the fields are duplicated).

I had to remove several derive attributes, because Handlebars doesn't
implement Clone, PartialEq or Debug. That's why I modified some builders
as well.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>