Reactor: use higher-ranked lifetimes and SharedData
This is the first patch which introduces higher-ranked lifetimes in this
library. I split a bigger change into smaller commits, so the code may
not be compiled through these changes.

The code cannot work with ordinary lifeties, because they would introduce
a lifetime bound on the Reactor implementation. That lifetime parameter
is substituted when the implementation is instantiated, so we get a
lifetime error...

Signed-off-by: Tibor Benke <ihrwein@gmail.com>