Correlator: add support for external handler data
When the correlator receives an event from the reactor, it executes
an event handler.

There are two event types: Alert and Shutdown. The former should be handled
by the user of the correlator, the latter one is an internal event, should
be handler by the Correlator itself.

There are various reasons to pass some data to the alert handler. A T type
without any bounds should provide enough flexibility, but there is a price
tag attached to this flexibility: the Correlator type become generic over
an unknown T type.

The users can register (set_alert_handler()) any AlertHandler type.

push_message() doesn't try to process the incoming messages from the reactor,
it should be called manually before push_message(). Explicit over implicit.

handle_event() will handle the incoming event from the reactor. It borrows
the external data for the alert handler.

stop_dispatcher() also borrows the external data for the handler.

These changes come with a great simplification in the Correlator's alert
handler's logic.

We don't need Rc and RefCell types in this crate any more :)