syslog-ng-common: prevent LogTemplate to use a dead GlobalConfig reference
Let's tie the cfg's lifetime to the LogTemplate struct by introducing
an internal marker field. It has sero size, but the compiler will look
at the LogTemplate struct as it has a reference to a GlobalConfig instance.

Note, that we would end up with solution even if we don't use raw pointers.
Rust doesn't track the lifetimes of raw pointers so we need the PhantomData.
If we store a reference directly, we don't need to PhantomData, but the
references are still required.

I tried to add compile-fail tests, but they are not-well documented.

Signed-off-by: Tibor Benke <tibor.benke@balabit.com>