correlation: simplify Correlator
The original intent was to run the correlation logic in its own thread, but
we faced one serious limitation: even if the correlator engine has its own
thread, the generated messages can be handled in the original thread (
which uses the Correlator type). So there was no gain in the plus thread,
just added complexity.

This commit greatly simplifies the Correlator type and this way prepares
the removal os several other types.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>