correlation-parser: refactor Timer
We can wrap it into an Arc<>, so we don't need it to be Clone.
That simplifies its internals, as there will be only one owner of it.
This means, we can store a JoinHandle directly.

During dropping, we send an event to the timer thread to start. When it
receives this event (StopEvent), it breaks out from an inner loop and
cease to exists.

When the joinhandle is dropped, it automatically detaches the thread if
it's not joined.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>