syslog-ng-common: add init() and deinit() methods to Parser trait
The init() method can be called in two contexts:
* right after the parser is built from a parser builder (new(), some
set_option() calls and an init())
* the parser was deinitialized with a deinit() call and we would like
  to use the parser again

init() and deinit() are always in pairs. They can be used to modify the
internal state of the parser when a reload event is received or the
parser is freshly built.

A common use-case is as follows:
* the parser builder creates a parser
* init() is called on the parser, which enables it setup and start timers
* there are some parse() calls
* deinit() is called, which stops the timers (and a future init() would
  start them again)

The two methods have default implementations, as most parsers are
stateless, so they don't need this much control.

Signed-off-by: Tibor Benke <tibor.benke@balabit.com>