syslog-ng-common: add commit_suicide() function
It calls abort().

We can use this functions to abort syslog-ng when we receive a fatal error
(like integer overflow, out-of-bound access to arrays). We catch
these errors on the Rust - C FFI boundary, but there is no guarantee that
the Rust parsers are in a valid state. Therefore, it's best to abort the
whole process.

We can't use panic! to abort the current process, as it causes a stack
unwinding which shouldn't step through the Rust-C boundary. That's why
we call libc's abort().

Signed-off-by: Tibor Benke <tibor.benke@balabit.com>